Do the necessary to finish the project
There was an issue previously with aligning the contents of the hero &
header... it was easily fixed today (really, where was my brain?!)

Along time was spent on trying to get CSS to automatically resize the
image in the hero to make it have the same dimensions as the contents
of the text block next to it... only to discover that this is probably
only achievable with javascript... solved the problem by substituting
the squared image by a rectangular one then adjusting the aspect ratio
and setting dimensions in pixels instead of percentages...

At least the project is done... albeit with silly text and zero
originality! no worries, if i ever choose to use it as a personal
landing page, of course i would change it drastically, for now let it
be just a boilerplate sitting in my chest of repositories!